+++ scg/expr/scene/scene.go
@@ -9,14 +9,14 @@ type Scene struct {
 	Text        Text              `yaml:"text" json:"text" xml:"text"`
 	NextScenes  []string          `yaml:"nextScenes,omitempty" json:"next_scenes,omitempty" xml:"nextScenes,omitempty"`
 	IsInfoScene bool              `yaml:"isInfoScene,omitempty" json:"is_info_scene,omitempty" xml:"isInfoScene,omitempty"`
-	Matchers    []Matcher         `yaml:"matchers,omitempty" json:"matchers,omitempty" xml:"matchers,omitempty"`
+	Matchers    []string          `yaml:"matchers,omitempty" json:"matchers,omitempty" xml:"matchers,omitempty"`
 	Error       Error             `yaml:"error,omitempty" json:"error,omitempty" xml:"error,omitempty"`
 	Buttons     map[string]Button `yaml:"buttons,omitempty" json:"buttons,omitempty" xml:"buttons,omitempty"`
 }
 
-func (s *Scene) IsValid() (bool, error) {
+func (s *Scene) IsValid(userMatchers map[string]Matcher) (bool, error) {
 
-	is, err := s.isMatchersValid()
+	is, err := s.isMatchersValid(userMatchers)
 
 	if !is {
 		return is, err
@@ -31,15 +31,16 @@ func (s *Scene) IsValid() (bool, error) {
 	return s.Text.IsValid()
 }
 
-func (s *Scene) isMatchersValid() (bool, error) {
+func (s *Scene) isMatchersValid(userMatchers map[string]Matcher) (bool, error) {
 	err := error(nil)
 	for _, matcher := range s.Matchers {
-		if matcher.IsDefaultMatcher() {
-			name, _ := matcher.GetStandardMatcher()
-			if !matchers.IsCorrectNameOfMather(name) {
-				err = errorNotSupportedMatherType(name)
-				break
-			}
+		if matchers.IsCorrectNameOfMather(matcher) {
+			continue
+		}
+
+		if _, is := userMatchers[matcher]; !is {
+			err = errorNotSupportedMatherType(matcher)
+			break
 		}
 	}