+++ scg/script/codegen/mathers.go
@@ -30,19 +30,32 @@ func generateMatchers(_ string, rootDir string, scriptInfo expr.ScriptInfo) *cod
 	regexMatchers := make([]scene.RegexMatcher, 0)
 	selectsMatchers := make([]scene.SelectMatcher, 0)
 
-	for _, sc := range scriptInfo.Script {
-		for _, matcher := range sc.Matchers {
-			if matcher.IsRegexMatcher() {
-				mc, _ := matcher.GetRegexMatcher()
-				regexMatchers = append(regexMatchers, *mc)
+	for name, matcher := range scriptInfo.UserMatchers {
+		if matcher.IsRegexMatcher() {
+			mc, _ := matcher.GetRegexMatcher()
+			if mc.Name == "" {
+				mc.Name = name
 			}
+			regexMatchers = append(regexMatchers, *mc)
+		}
 
-			if matcher.IsSelectMatcher() {
-				mc, _ := matcher.GetSelectsMatcher()
-				selectsMatchers = append(selectsMatchers, *mc)
+		if matcher.IsSelectMatcher() {
+			mc, _ := matcher.GetSelectsMatcher()
+			if mc.Name == "" {
+				mc.Name = name
 			}
+			selectsMatchers = append(selectsMatchers, *mc)
 		}
 	}
+
+	sections = append(sections, &codegen.SectionTemplate{
+		Name:   "regex-consts-matchers",
+		Source: regexConstStructT,
+		Data:   regexMatchers,
+		FuncMap: map[string]interface{}{
+			"ToTitle": codegen.ToTitle,
+		},
+	})
 
 	sections = append(sections, &codegen.SectionTemplate{
 		Name:   "regex-matchers",
@@ -74,10 +87,20 @@ func generateMatchers(_ string, rootDir string, scriptInfo expr.ScriptInfo) *cod
 	return &codegen.File{Path: fpath, SectionTemplates: sections, IsUpdatable: true}
 }
 
+const regexConstStructT = `
+{{ if . }}// name matched string for RegexMatchers
+const (
+{{range .}}
+	{{ ToTitle .Name }}MatchedString = "{{ .NameMatched }}"
+{{end}}
+) {{end}}
+
+`
+
 const regexMatchersStructT = `{{ if . }}// RegexMatchers
 var (
 {{range .}}
-		{{ ToTitle .Name }}Matcher = matchers.NewRegexMather("{{ .Regex }}")
+		{{ ToTitle .Name }}Matcher = matchers.NewRegexMather("{{ .Regex }}", {{ ToTitle .Name }}MatchedString)
 {{end}}
 ) {{end}}
 `
@@ -86,7 +109,7 @@ const selectsConstStructT = `
 {{ if . }}// replace string for SelectsMatchers
 const (
 {{range .}}
-	{{ ToTitle .Name }}MatcherString = "{{ .ReplaceMessage }}"
+	{{ ToTitle .Name }}MatchedString = "{{ .ReplaceMessage }}"
 {{end}}
 ) {{end}}
 
@@ -99,7 +122,7 @@ var (
 	[]string{ {{ range .Selects }}
 		"{{ . }}",{{end}}
 	},
-	{{ ToTitle .Name }}MatcherString,
+	{{ ToTitle .Name }}MatchedString,
 	)
 {{end}}
 ) {{end}}